// client/src/features/inventory-item/types/inventoryItemEnum.ts
export enum InventoryItemCategoryEnum {
  ENGINE = 1,
  TRANSMISSION = 2,
  BRAKES = 3,
  TIRES = 4,
  ELECTRICAL = 5,
  BODY = 6,
  INTERIOR = 7,
  FLUIDS = 8,
  FILTERS = 9,
}

export enum InventoryItemUnitCostMeasurementUnitEnum {
  Unit = 1,
  Litre = 2,
  Gram = 3,
  Kilogram = 4,
  Metre = 5,
  SquareMetre = 6,
  CubicMetre = 7,
  Box = 8,
}